Modify to Double instead of Float and adapt doc

+++ src/main/java/com/danielpsf/labs/field/statistic/domain/request/StatisticRequest.java
@@ -9,6 +9,6 @@ import java.util.Date;
 @Builder
 public class StatisticRequest {
 
-    private Float vegetation;
+    private Double vegetation;
     private Date occurrenceAt;
 }
